Renamed to curvehooks_impl to base_impl
