+++ src/lib.rs
@@ -17,8 +17,8 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![doc = include_str!("../README.md")]
 
+pub mod base_impl;
 mod constants;
-pub mod curvehooks_impl;
 pub mod errors;
 pub mod key;
 pub mod proof;
